switch to using the quivr type as precovery output
