+++ tests/test_precovery_search.py
@@ -27,6 +27,8 @@ def test_precover(precovery_db, sample_orbits):
 
     # Do the search. We should find the three observations we inserted.
     matches, misses = list(precovery_db.precover(orbit))
+    matches = matches.to_dataclass()
+    misses = misses.to_frame_candidates()
     assert len(matches) == 3
     assert len(misses) == 0
 
@@ -58,6 +60,8 @@ def test_precover_dataset_filter(precovery_db, sample_orbits):
     # Do the search with no dataset filters. We should find all six
     # observations we inserted.
     matches, misses = list(precovery_db.precover(orbit))
+    matches = matches.to_dataclass()
+    misses = misses.to_frame_candidates()
     assert len(matches) == 6
 
     have_ids = set(r.observation_id for r in matches)
@@ -67,6 +71,8 @@ def test_precover_dataset_filter(precovery_db, sample_orbits):
     # Now repeat the search, but filter to just one dataset. We should
     # only find that dataset's observations.
     matches, misses = list(precovery_db.precover(orbit, datasets={ds1_id}))
+    matches = matches.to_dataclass()
+    misses = misses.to_frame_candidates()
     assert len(matches) == 3
 
     have_ids = set(r.observation_id for r in matches)
